Update curry_exercises.js
The curry_exercises.js file was missing some help text in a comment which was in the book.  I found the help text was useful to understand what the problem was asking.  Original comment just said "use slice to define a function "take" that takes n elements" and it was unclear whether it involved an array.